+++ src/app/app-routing.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
+
 import { AuthGuard } from './guard/guard.guard';
 import {HomeComponent} from 'src/app/components/home/home.component'
 import {LoginComponent} from 'src/app/components/login/login.component'