+++ src/app/components/login/login.component.spec.ts
@@ -11,7 +11,6 @@ import { BotaoComponent } from '../botao/botao.component';
 import { AuthService } from '../../services/auth/auth.service';
 import { BotaoTemaComponent } from '../botao-tema/botao-tema.component';
 import { MensagensService } from 'src/app/services/mensagens/mensagens.service';
-import { LogoParceiroComponent } from '../logo-parceiro/logo-parceiro.component'; 
 import { InputSenhaComponent } from '../input-senha/input-senha.component';
 
 
@@ -42,7 +41,6 @@ describe('LoginComponent', () => {
         InputComponent,
         BotaoTemaComponent,
         InputSenhaComponent,
-        LogoParceiroComponent,
       ],
       providers: [
         { provide: AuthService, useValue: authServiceMock },