+++ src/app/components/redefinir-senha/redefinir-senha.component.spec.ts
@@ -8,7 +8,6 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { RedefinirSenhaComponent } from './redefinir-senha.component';
 import { BotaoTemaComponent } from '../botao-tema/botao-tema.component';
 import { InputSenhaComponent } from '../input-senha/input-senha.component';
-import { LogoParceiroComponent } from '../logo-parceiro/logo-parceiro.component';
 
 
 describe('RedefinirSenhaComponent', () => {
@@ -29,7 +28,6 @@ describe('RedefinirSenhaComponent', () => {
         BotaoComponent,
         InputComponent,
         InputSenhaComponent,
-        LogoParceiroComponent,
         BotaoTemaComponent,
         RedefinirSenhaComponent,
       ],