+++ src/app/components/registro/registro.component.ts
@@ -12,8 +12,8 @@ import { MensagensService } from 'src/app/services/mensagens/mensagens.service';
 })
 export class RegistroComponent {
   public imgSrc?: string;
-  private imgTemaClaro: string = 'assets/img/logo-parceiro-vertical-preto-1.png';
-  private imgTemaEscuro: string = 'assets/img/imgogo-parceiro-vertical-branco-1.png';
+  private imgTemaClaro: string = '';
+  private imgTemaEscuro: string = '';
   constructor(
     private router: Router,
     private authService: AuthService,